chore: Move `BaseSkeleton` into own `base.py`-module
