+++ src/viur/core/skeleton/__init__.py
@@ -2,8 +2,9 @@ import logging
 import warnings
 
 from .adapter import DatabaseAdapter, ViurTagsSearchAdapter
+from .base import BaseSkeleton
 from .instance import SkeletonInstance
-from .meta import ABSTRACT_SKEL_CLS_SUFFIX, BaseSkeleton, MetaBaseSkel, MetaSkel, Skeleton_Cls
+from .meta import ABSTRACT_SKEL_CLS_SUFFIX, MetaBaseSkel, MetaSkel, Skeleton_Cls
 from .relskel import RefSkel, RelSkel
 from .skeleton import SeoKeyBone, Skeleton, _UNDEFINED_KINDNAME
 from .tasks import SkelIterTask, SkeletonMaintenanceTask, update_relations
